fix: make sure the interactive printers can cleanup after Ctrl+C
After canceling the interactive printers with Ctrl+C, `os.Exit` was
directly called in the keyboard listener. Because of this the
`defer` statements of the surrounding function were not called
(for example `cursor.Show`) and the keyboard listener
was still running.
This resulted in a broken user "terminal" (no Cursor, no Ctrl+C)