fix(prefix): PrefixPrinter now uses updated defaultWriter
When calling pterm.SetDefaultOuput(writer), the prefix printers
initialized by pterm did not use the new default.

This commit changes this behaviour, and the prefix printers will now use
whatever is set as the defaultWriter unless a specific writer has been
set using the WithWriter(writer) method on the PrefixPrinter, in which
case it will use that printer.

Closes #751